Fixed using %*s format in nxt_printf().

+++ njs/njs_disassembler.c
@@ -420,7 +420,7 @@ njs_disassemble(u_char *start, u_char *end)
                     code3 = (njs_vmcode_3addr_t *) p;
 
                     nxt_printf("%05uz %*s  %04Xz %04Xz %04Xz\n",
-                               p - start, (int) name->length, name->start,
+                               p - start, name->length, name->start,
                                (size_t) code3->dst, (size_t) code3->src1,
                                (size_t) code3->src2);
 
@@ -428,14 +428,14 @@ njs_disassemble(u_char *start, u_char *end)
                     code2 = (njs_vmcode_2addr_t *) p;
 
                     nxt_printf("%05uz %*s  %04Xz %04Xz\n",
-                               p - start, (int) name->length, name->start,
+                               p - start, name->length, name->start,
                                (size_t) code2->dst, (size_t) code2->src);
 
                 } else if (code_name->size == sizeof(njs_vmcode_1addr_t)) {
                     code1 = (njs_vmcode_1addr_t *) p;
 
                     nxt_printf("%05uz %*s  %04Xz\n",
-                               p - start, (int) name->length, name->start,
+                               p - start, name->length, name->start,
                                (size_t) code1->index);
                 }
